# apps/frontend/src/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/gestion-despacho": {
      "post": {
        "description": "Inserta un nuevo despacho en la base de datos",
        "operationId": "GestionDespachoController_insertarDespacho",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Despacho a insertar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDespachoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Despacho insertado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Insertar un despacho",
        "tags": [
          "GestionDespacho"
        ]
      }
    },
    "/api/gestion-despacho/rechazo": {
      "post": {
        "description": "Registra un nuevo rechazo de despacho en la base de datos",
        "operationId": "GestionDespachoController_registrarRechazoDespacho",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Rechazo a registrar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRechazoDespachoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Rechazo de despacho registrado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Registrar rechazo de despacho",
        "tags": [
          "GestionDespacho"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Crea un nuevo usuario con contraseña encriptada",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Datos del usuario a registrar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Registrar nuevo usuario",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Autentica un usuario y devuelve un token JWT que contiene el ID del usuario",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Credenciales de acceso",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso - Token JWT generado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/LoginResponseDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "description": "Obtiene la información del usuario autenticado (solo admin, supervisor, operador y monitor)",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Perfil del usuario obtenido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserProfileDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado - Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtener perfil del usuario",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/users": {
      "get": {
        "description": "Obtiene los primeros 10 usuarios con opción de búsqueda por nombre",
        "operationId": "AuthController_searchUsers",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Término de búsqueda para filtrar por nombre de usuario",
            "schema": {
              "example": "admin",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Límite de usuarios a retornar (por defecto 10)",
            "schema": {
              "example": 10,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuarios obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserSearchDto"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Acceso denegado - Rol insuficiente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar usuarios",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/control-entrega/ordenes-pendientes/{idUsuario}": {
      "get": {
        "description": "Obtiene todas las órdenes de entrega pendientes para un usuario específico",
        "operationId": "OrdenesPendientesController_obtenerOrdenesPendientes",
        "parameters": [
          {
            "name": "idUsuario",
            "required": true,
            "in": "path",
            "description": "ID único del usuario",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Órdenes pendientes obtenidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Operación exitosa"
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrdenesPendientesResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID de usuario inválido o formato incorrecto"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Obtener órdenes pendientes de un usuario",
        "tags": [
          "Control de Entrega - Órdenes Pendientes"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Transporte",
    "description": "API para el sistema de monitoreo de transporte",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Ingresa tu token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "EmptyResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "Estado de la respuesta",
            "example": true,
            "nullable": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje de la respuesta",
            "example": "Operación exitosa",
            "nullable": false
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "boolean",
            "description": "Estado de la respuesta",
            "example": false,
            "nullable": false
          },
          "message": {
            "type": "string",
            "description": "Mensaje de error",
            "example": "Error interno del servidor",
            "nullable": false
          }
        },
        "required": [
          "status",
          "message"
        ]
      },
...
      "CreateDespachoDto": {
        "type": "object",
        "properties": {
          "id_resultado_condicion_despacho": {
            "type": "string",
            "description": "ID del resultado de la condición del despacho"
          },
          "id_usuario": {
            "type": "string",
            "description": "ID del usuario que crea el despacho"
          },
          "estado_recorrido": {
            "type": "string",
            "description": "Estado del recorrido",
            "enum": [
              "en_ruta",
              "completada",
              "cancelada"
            ],
            "default": "en_ruta"
          },
          "observaciones": {
            "type": "string",
            "description": "Observaciones del despacho"
          }
        },
        "required": [
          "id_resultado_condicion_despacho",
          "id_usuario",
          "estado_recorrido"
        ]
      },
      "CreateRechazoDespachoDto": {
        "type": "object",
        "properties": {
          "id_resultado_condicion_despacho": {
            "type": "string",
            "description": "ID del resultado de la condición del despacho"
          },
          "id_usuario": {
            "type": "string",
            "description": "ID del usuario que crea el rechazo de despacho"
          },
          "motivo_rechazo": {
            "type": "string",
            "description": "Motivo del rechazo",
            "maxLength": 200
          },
          "tipo_accion": {
            "type": "string",
            "description": "Tipo de acción a realizar",
            "enum": [
              "reasignar_vehiculo",
              "reasignar_conductor",
              "mover_carga",
              "reprogramar",
              "cancelar"
            ]
          },
          "estado_accion": {
            "type": "string",
            "description": "Estado actual de la acción",
            "enum": [
              "pendiente",
              "en_proceso",
              "completada",
              "cancelada"
            ],
            "default": "pendiente"
          }
        },
        "required": [
          "id_resultado_condicion_despacho",
          "id_usuario",
          "motivo_rechazo",
          "tipo_accion"
        ]
      },
      "UserSearchDto": {
        "type": "object",
        "properties": {
          "id_usuario": {
            "type": "string",
            "description": "ID del usuario",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nombre_usuario": {
            "type": "string",
            "description": "Nombre de usuario",
            "example": "admin"
          },
          "rol": {
            "type": "string",
            "description": "Rol del usuario",
            "enum": [
              "operador",
              "admin",
              "supervisor",
              "monitor"
            ],
            "example": "admin"
          }
        },
        "required": [
          "id_usuario",
          "nombre_usuario",
          "rol"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "nombre_usuario": {
            "type": "string",
            "description": "Nombre de usuario",
            "example": "nuevo_usuario"
          },
          "contrasena": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "password123"
          },
          "rol": {
            "type": "string",
            "description": "Rol del usuario",
            "enum": [
              "operador",
              "admin",
              "supervisor",
              "monitor"
            ],
            "example": "operador"
          }
        },
        "required": [
          "nombre_usuario",
          "contrasena",
          "rol"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Token JWT generado",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "token"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "nombre_usuario": {
            "type": "string",
            "description": "Nombre de usuario",
            "example": "admin"
          },
          "contrasena": {
            "type": "string",
            "description": "Contraseña del usuario",
            "example": "password123"
          }
        },
        "required": [
          "nombre_usuario",
          "contrasena"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "nombre_usuario": {
            "type": "string",
            "description": "Nombre de usuario",
            "example": "admin"
          },
          "rol": {
            "type": "string",
            "description": "Rol del usuario",
            "enum": [
              "operador",
              "admin",
              "supervisor",
              "monitor"
            ],
            "example": "admin"
          }
        },
        "required": [
          "nombre_usuario",
          "rol"
        ]
      },
      "OrdenPendienteDto": {
        "type": "object",
        "properties": {
          "id_entrega": {
            "type": "string",
            "description": "ID único de la entrega",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "nombre_punto": {
            "type": "string",
            "description": "Nombre del punto de entrega",
            "example": "Restaurante El Buen Sabor"
          },
          "coordenada_punto": {
            "type": "string",
            "description": "Coordenadas del punto de entrega",
            "example": "(12.3456,-78.9012)"
          },
          "fecha_entrega": {
            "type": "string",
            "description": "Fecha de entrega programada",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "total_subentregas": {
            "type": "number",
            "description": "Total de subentregas asociadas",
            "example": 3
          }
        },
        "required": [
          "id_entrega",
          "nombre_punto",
          "coordenada_punto",
          "fecha_entrega",
          "total_subentregas"
        ]
      },
      "OrdenesPendientesResponseDto": {
        "type": "object",
        "properties": {
          "ordenes": {
            "description": "Lista de órdenes pendientes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenPendienteDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total de órdenes pendientes",
            "example": 5
          }
        },
        "required": [
          "ordenes",
          "total"
        ]
      },
...
    }
  }
}